+++ src/actions/index.js
@@ -1,0 +1,8 @@
+import {ADD_LABEL} from './types'
+
+export function addLabel(label){
+	return {
+		type: ADD_LABEL,
+		payload: label
+	}
+}