+++ src/components/map/index.js
@@ -1,4 +1,5 @@
 import React, { Component } from 'react';
+import { connect } from 'react-redux'
 
 import styles from './styles.scss'
 
@@ -9,13 +10,16 @@ class Map extends Component {
   	this.state={
     	markers:[]
     }
+
+    this.map = this.map.bind(this);
+    this.placeMarker = this.placeMarker.bind(this);
   }
 
   componentDidMount () {
     this.map();
     this.map = this.map.bind(this);
-    this.placeMarker = this.placeMarker.bind(this);
-    this.renderLabel = this.renderLabel.bind(this);
+    console.log('PROPS', this.props)
+
   }
 
   map() {
@@ -27,31 +31,55 @@ class Map extends Component {
       },
       zoom: 16
     })
+
+    this.renderMarkers(this.props.markers, map);
+
     map.addListener('click', function(e) {
 	    placeMarker(e.latLng, map);
 	});
   }
 
-  placeMarker(location, map){
-  	console.log(location, map)
-
-  	let infoContent = 'TEST'
-
-  	let infoWindow = new google.maps.InfoWindow({
-  		content: infoContent
-  	})
-
-    let marker = new google.maps.Marker({
+  marker(location, map){
+    return new google.maps.Marker({
         position: location, 
         map: map,
-        title: 'test'
     })
+  }
+
+  infoWindow(infoContent){
+    return new google.maps.InfoWindow({
+      content: infoContent
+    })
+  }
+
+  infoListener(marker, infoWindow){
     marker.addListener('click', function(){
-    	infoWindow.open(map, marker)
+      infoWindow.open(map, marker)
     })
+  }
+  placeMarker(location, map){
+  	console.log(location)
+
+  	let infoWindow = this.infoWindow('TEST')
+
+    let marker = this.marker(location,map);
+
+    console.log('MARKER LAT', marker.position.lat())
+    console.log('MARKER LAT', marker.position.lng())
+    let markerListener = this.infoListener(marker, infoWindow);
     map.panTo(location) 	
   }
 
+  renderMarkers(markers, map){
+    markers.map((marker, index) =>{
+      let {title, latitude, longitude} = marker;
+      let location = {lat: latitude, lng: longitude};
+      let newMarker = this.marker(location, map)
+      let infoWindow = this.infoWindow(title)
+      let markerListener = this.infoListener(newMarker, infoWindow);
+    })
+  }
+  
   render() {
     return (
       <div>
@@ -61,4 +89,8 @@ class Map extends Component {
   }
 }
 
-export default Map;
+function mapStateToProps({markers}){
+  return{markers}
+}
+
+export default connect(mapStateToProps)(Map);